+++ src/colifer/constants.py
@@ -1,0 +1,3 @@
+"""Constants for the whole project"""
+
+LOGGING_FORMAT = '[%(levelname)s] %(asctime)s %(name)s:%(lineno)d - %(message)s'